backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
pydantic
requests
numpy
chromadb
tinydb